@@ -102,9 +102,9 @@
     "qr-code-styling": "^1.8.4",
     "qr-scanner": "^1.4.2",
     "raw-loader": "^4.0.2",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-component-transition": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-highlight": "^0.14.0",
     "react-hook-form": "^7.53.2",
     "react-infinite-scroll-component": "^6.1.0",
